requirements.txt
Django>=4.2
gunicorn>=21.0
Pillow==9.5.0
whitenoise>=6.6
psycopg2-binary>=2.9
python-dotenv>=1.0